chores(docs): Move example from markdown to tests (#89)
It's required because:
- package users can use godoc.org, pkg.go.dev for exploring
the package documentation
- testable examples can be more in sync with actual code

This commit moves examples from the plain markdown to testable
examples. It allows to test examples during test phase and have
them more in sync with actual code. Also the examples can be
shown by documentation servers and in IDEs and make work with
package much more user-friendly as not everyone looks into
README.md at repo's root.

Related to #68